Put misfiled types where they belong, and split the settings file
Housekeeping from the extraction audit. Every item is a move or a file split;
no behaviour changes, 1759/1759 green.

Misfiled types:
- ApplicationUpdateService -> Ui/. It is a WinForms controller (Form owner,
  MessageBox, Application.Exit, ApplicationUpdateDialog) that was sitting in
  the persistence folder.
- OverlayDialogControls -> Ui/. Control factory, not overlay logic.
- VirtualCrossoverSheet + VirtualCrossoverSheetPdf -> Tools/Sheets/, next to
  PdfSheet and TuningSheetPdf.
- InputLevelMeterSnapshot -> Measurements/. The type defines part of the .ir
  and history schema; Ui/Controls was the wrong owner for a persisted shape.
- LiveSpectrumOptions / NoiseColor / AveragingSpeed -> Measurements/, out of
  namespace Resonalyze.Options. Options/ holds the WinForms settings panels,
  but NoiseMeasurement takes this object by reference and reads it on the
  analysis path, so it is measurement state.
- WaterfallGenerateOptions + WaterfallMode -> their own file in Measurements/.
  They are a persisted settings schema and were living in WaterfallSeries.cs
  beside an OxyPlot XYAxisSeries.

Mode moves out of Form1.cs into ModeSwitching/Mode.cs. It is not a shell
detail: overlays, plotting, curve tags and the settings file are all typed on
it. Its doc comment now records that the member NAMES are persisted —
OverlayFile serializes them and uses mode.ToString() as the on-disk directory —
so a rename orphans saved overlays.

MeasurementSettingsFile (1047 lines) becomes three partials by the job each
part does: the file itself keeps load/save/backup and the Apply/Capture entry
points; .Schema.cs holds the per-mode nested settings classes; and
.DeviceResolution.cs holds the normalizers, which are not mapping at all —
they enumerate real hardware through WindowsAudioEndpointService,
AsioDeviceCatalog and AudioDeviceCatalog to decide whether a stored device
still exists.

MeasurementHistorySnapshotMetadata.BuildToolTipText becomes
Ui/MeasurementHistoryToolTip.Build. String formatting and ToolTipTextWrapper
had no business on a persisted schema record; the three call sites move with
it. StartupAudioWarmup and MicrophoneCalibrationService move to Shell/, which
is what their own doc comments say they are — Form1 decomposition leftovers.
