+++ source/Measurements/InputLevelMeterSnapshot.cs
