+++ source/Ui/ApplicationUpdateService.cs
