+++ source/Ui/OverlayDialogControls.cs
